rename logs into log_style
